Instead of only including PSM(s) with the best score for a given scan, all PSMs are now included.
Added a filterable annotation called "rank", which is 1 for the top
scoring PSM(s) for a given scan. (All PSMs tied for the best score have
rank 1). All PSMs tied for the second best score have rank = 1 + the
number of PSMs tied for rank 1. And so on.

Made rank <= 1 a default filter.

PSM rank is now shown by default.